requirements.txt
Django>=5.0
Pillow>=10.0.0
django-import-export>=4.0
psycopg[binary]>=3.0.0